Discrete Lehmann Representation

Lehmann.DLR — Module

discrete Lehmann representation for imaginary-time/Matsubara-freqeuncy correlator

Lehmann.DLR.DLRGrid — Type

struct DLRGrid

DLR grids for imaginary-time/Matsubara frequency correlators

#Members:

type: symbol :fermi, :corr, :acorr

Euv : the UV energy scale of the spectral density

β : inverse temeprature

Λ: cutoff = UV Energy scale of the spectral density * inverse temperature

rtol: tolerance absolute error

size : number of DLR basis

ω : selected representative real-frequency grid

n : selected representative Matsubara-frequency grid (integer)

ωn : (2n+1)π/β

τ : selected representative imaginary-time grid

Lehmann.DLR.barycheb — Method

function barycheb(n, x, f, wc, xc)

Barycentric Lagrange interpolation at Chebyshev nodes Reference: Berrut, J.P. and Trefethen, L.N., 2004. Barycentric lagrange interpolation. SIAM review, 46(3), pp.501-517.

Arguments

n: order of the Chebyshev interpolation

x: coordinate to interpolate

f: array of size n, function at the Chebyshev nodes

wc: array of size n, Barycentric Lagrange interpolation weights

xc: array of size n, coordinates of Chebyshev nodes

Returns

Interpolation result

Lehmann.DLR.barychebinit — Method

barychebinit(n)

Get Chebyshev nodes of first kind and corresponding barycentric Lagrange interpolation weights. Reference: Berrut, J.P. and Trefethen, L.N., 2004. Barycentric lagrange interpolation. SIAM review, 46(3), pp.501-517.

Arguments

n: order of the Chebyshev interpolation

Returns

Chebyshev nodes

Barycentric Lagrange interpolation weights

Lehmann.DLR.dlr — Method

function dlr(type, Λ, rtol) Construct discrete Lehmann representation

#Arguments:

type: type of kernel, :fermi, :boson

Λ: cutoff = UV Energy scale of the spectral density * inverse temperature

rtol: tolerance absolute error

Lehmann.DLR.dlr2matfreq — Function

function dlr2matfreq(type, dlrcoeff, dlrGrid::DLRGrid, nGrid, β=1.0; axis=1)

DLR representation to Matsubara-frequency representation

#Members:

type: symbol :fermi, :corr, :acorr

dlrcoeff : DLR coefficients

dlrGrid : DLRGrid

nGrid : expected fine Matsubara-freqeuncy grids (integer)

axis: Matsubara-frequency axis in the data dlrcoeff

rtol: tolerance absolute error

Lehmann.DLR.dlr2tau — Method

function dlr2tau(type, dlrcoeff, dlrGrid::DLRGrid, τGrid; axis=1)

DLR representation to imaginary-time representation

#Members:

type: symbol :fermi, :corr, :acorr

dlrcoeff : DLR coefficients

dlrGrid : DLRGrid

τGrid : expected fine imaginary-time grids ∈ (0, β]

axis: imaginary-time axis in the data dlrcoeff

rtol: tolerance absolute error

Lehmann.DLR.kernalDiscretization — Method

kernalDiscretization(type, Nτ, Nω, Λ, rtol)

Discretize kernel K(tau,omega) on composite Chebyshev fine grids for τ and ω. Generate a panels and grids for τ and ω.

#Arguments:

type: :fermi, :bose or :corr

Dτ: Chebyshev degree in each τ panel

Dω: Chebyshev degree in each ω panel

Λ: cutoff = UV Energy scale of the spectral density * inverse temperature

rtol: tolerance relative error

#Returns

τPanel: panel break points for τ, exponentially get dense near 0⁺ and 1⁻

ωPanel: panel break points for ω, get exponentially dense near 0⁻ and 0⁺

τGrid: tau fine grid points on (0,1)

ωGrid: omega fine grid points on (-Λ, Λ)

kernel: K(tau,omega) on fine grid

err: Error of composite Chebyshev interpolant of K(tau,omega). err(1) is ~= max relative L^inf error in tau over all omega in fine grid. err(2) is ~= max L^inf error in omega over all tau in fine grid.

Lehmann.DLR.matfreq2dlr — Method

function matfreq2dlr(type, green, dlrGrid::DLRGrid; axis=1, rtol=1e-12)

Matsubara-frequency representation to DLR representation

#Members:

type: symbol :fermi, :corr, :acorr

green : green's function in Matsubara-frequency domain

axis: the Matsubara-frequency axis in the data green

rtol: tolerance absolute error

Lehmann.DLR.matfreq2tau — Method

function matfreq2tau(type, green, dlrGrid, τGrid; axis=1, rtol=1e-12)

Fourier transform from Matsubara-frequency to imaginary-time using the DLR representation

#Members:

type: symbol :fermi, :corr, :acorr

green : green's function in Matsubara-freqeuncy repsentation

dlrGrid : DLRGrid

τGrid : expected fine imaginary-time grids

axis: Matsubara-frequency axis in the data green

rtol: tolerance absolute error

Lehmann.DLR.tau2dlr — Method

function tau2dlr(type, green, dlrGrid::DLRGrid; axis=1, rtol=1e-12)

imaginary-time domain to DLR representation

#Members:

type: symbol :fermi, :corr, :acorr

green : green's function in imaginary-time domain

axis: the imaginary-time axis in the data green

rtol: tolerance absolute error

Lehmann.DLR.tau2matfreq — Method

function tau2matfreq(type, green, dlrGrid, nGrid; axis=1, rtol=1e-12)

Fourier transform from imaginary-time to Matsubara-frequency using the DLR representation

#Members:

type: symbol :fermi, :corr, :acorr

green : green's function in imaginary-time domain

dlrGrid : DLRGrid

nGrid : expected fine Matsubara-freqeuncy grids (integer)

axis: the imaginary-time axis in the data green

rtol: tolerance absolute error