Sample Green's Function Builder

This module provides subroutines to generate Green's function samples.

Lehmann.Sample.MultiPole — Function

MultiPole(β, isFermi::Bool, symmetry::Symbol, Grid, type::Symbol, poles, regularized::Bool = true)

Generate Green's function from a spectral density with delta peaks specified by the argument $poles$. Return the function on Grid and the systematic error.

#Arguments

β : inverse temperature

isFermi: is fermionic or bosonic

symmetry: particle-hole symmetric :ph, particle-hole antisymmetric :pha, or :none

Grid: grid to evalute on

type: imaginary-time with :τ, or Matsubara-frequency with :ωn

poles: a list of frequencies for the delta functions

regularized: use regularized bosonic kernel if symmetry = :none

Lehmann.Sample.SemiCircle — Function

SemiCircle(Euv, β, isFermi::Bool, symmetry::Symbol, Grid, type::Symbol, rtol = nothing, degree = 24, regularized::Bool = true)

Generate Green's function from a semicircle spectral density. Return the function on Grid and the systematic error.

#Arguments

Euv : ultraviolet energy cutoff

β : inverse temperature

isFermi: is fermionic or bosonic

symmetry: particle-hole symmetric :ph, particle-hole antisymmetric :pha, or :none

Grid: grid to evalute on

type: imaginary-time with :τ, or Matsubara-frequency with :ωn

rtol: accuracy to achieve

degree: polynomial degree for integral

regularized: use regularized bosonic kernel if symmetry = :none