Spectral functions

Lehmann.Spectral — Module

Spectral representation related functions

Lehmann.Spectral.boseEinstein — Method

boseEinstein(ω)

Compute the Fermi Dirac function. Assume $k_B T/\hbar=1$

\[f(ω) = 1/(e^{ωβ}-1)\]

Arguments

ω: frequency

β: the inverse temperature

Lehmann.Spectral.density — Method

density(type, ω, β=1.0)

Compute the imaginary-time kernel of different type. Assume $k_B T/\hbar=1$

Arguments

type: symbol :fermi, :bose

ω: energy

β: the inverse temperature

Lehmann.Spectral.fermiDirac — Method

fermiDirac(ω)

Compute the Fermi Dirac function. Assume $k_B T/\hbar=1$

\[f(ω) = 1/(e^{ωβ}+1)\]

Arguments

ω: frequency

β: the inverse temperature

Lehmann.Spectral.kernelAnormalCorrT — Method

kernelAnormalCorrT(τ, ω, β=1.0)

Compute the imaginary-time kernel for correlation function $⟨O(τ)O(0)⟩$. Machine accuracy ~eps(C) is guaranteed``

\[K(τ) = e^{-ω|τ|}-e^{-ω(β-|τ|)}\]

Arguments

τ: the imaginary time, must be (0, β]

ω: frequency, ω>=0

β = 1.0: the inverse temperature

Lehmann.Spectral.kernelAnormalCorrΩ — Method

kernelAnormalCorrΩ(n::Int, ω::T, β::T) where {T <: AbstractFloat}

Compute the Matsubara-frequency kernel for a anormalous fermionic correlator with particle-hole symmetry.

\[K(iω_n) = \frac{2ω}{ω^2+ω_n^2}(1+e^{-ωβ}),\]

where $ω_n=(2n+1)π/β$. The convention here is consist with the book "Quantum Many-particle Systems" by J. Negele and H. Orland, Page 95

Arguments

n: index of the fermionic Matsubara frequency

ω: energy

β: the inverse temperature

Lehmann.Spectral.kernelBoseT — Method

kernelBoseT(τ, ω, β=1.0)

Compute the imaginary-time bosonic kernel. Machine accuracy ~eps(g) is guaranteed``

\[g(τ>0) = e^{-ωτ}/(1-e^{-ωβ}), g(τ≤0) = -e^{-ωτ}/(1-e^{ωβ})\]

Arguments

τ: the imaginary time, must be (-β, β]

ω: frequency

β = 1.0: the inverse temperature

Lehmann.Spectral.kernelBoseΩ — Method

kernelBoseΩ(n::Int, ω::T, β::T) where {T <: AbstractFloat}

Compute the bosonic kernel with Matsubara frequency.

\[g(iω_n) = -1/(iω_n-ω),\]

where $ω_n=2nπ/β$. The convention here is consist with the book "Quantum Many-particle Systems" by J. Negele and H. Orland, Page 95

Arguments

n: index of the Matsubara frequency

ω: energy

β: the inverse temperature

Lehmann.Spectral.kernelCorrT — Method

kernelCorrT(τ, ω, β=1.0)

Compute the imaginary-time kernel for correlation function $⟨O(τ)O(0)⟩$. Machine accuracy ~eps(C) is guaranteed``

\[K(τ) = e^{-ω|τ|}+e^{-ω(β-|τ|)}\]

Arguments

τ: the imaginary time, must be (-β, β]

ω: frequency, ω>=0

β = 1.0: the inverse temperature

Lehmann.Spectral.kernelCorrΩ — Method

kernelCorrΩ(n::Int, ω::T, β::T) where {T <: AbstractFloat}

Compute the Matsubara-frequency kernel for a correlator $⟨O(τ)O(0)⟩_{iω_n}$.

\[K(iω_n) = \frac{2ω}{ω^2+ω_n^2}(1-e^{-ωβ}),\]

where $ω_n=2nπ/β$. The convention here is consist with the book "Quantum Many-particle Systems" by J. Negele and H. Orland, Page 95

Arguments

n: index of the Matsubara frequency

ω: energy

β: the inverse temperature

Lehmann.Spectral.kernelFermiT — Method

kernelFermiT(τ, ω, β=1.0)

Compute the imaginary-time fermionic kernel. Machine accuracy ~eps(g) is guaranteed``

\[g(τ>0) = e^{-ωτ}/(1+e^{-ωβ}), g(τ≤0) = -e^{-ωτ}/(1+e^{ωβ})\]

Arguments

τ: the imaginary time, must be (-β, β]

ω: frequency

β = 1.0: the inverse temperature

Lehmann.Spectral.kernelFermiΩ — Method

kernelFermiΩ(n::Int, ω::T, β::T) where {T <: AbstractFloat}

Compute the fermionic kernel with Matsubara frequency.

\[g(iω_n) = -1/(iω_n-ω),\]

where $ω_n=(2n+1)π/β$. The convention here is consist with the book "Quantum Many-particle Systems" by J. Negele and H. Orland, Page 95

Arguments

n: index of the Matsubara frequency

ω: energy

β: the inverse temperature

Lehmann.Spectral.kernelT — Method

kernelT(type::Symbol, τGrid::Vector{T}, ωGrid::Vector{T}, β::T=1.0) where {T<:AbstractFloat}

Compute kernel with given τ and ω grids.

Lehmann.Spectral.kernelT — Method

kernelT(type, τ, ω, β=1.0)

Compute the imaginary-time kernel of different type.

Arguments

type: symbol :fermi, :bose, :corr

τ: the imaginary time, must be (-β, β]

ω: frequency

β = 1.0: the inverse temperature

Lehmann.Spectral.kernelΩ — Method

kernelΩ(type, n, ω, β=1.0)

Compute the imaginary-time kernel of different type. Assume $k_B T/\hbar=1$

Arguments

type: symbol :fermi, :bose, :corr

n: index of the Matsubara frequency

ω: energy

β: the inverse temperature

Lehmann.Spectral.kernelΩ — Method

kernelΩ(type::Symbol, nGrid::Vector{Int}, ωGrid::Vector{T}, β::T=1.0) where {T<:AbstractFloat}

Compute kernel matrix with given ωn (integer!) and ω grids.